+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add package#is_empty method
 * Add package#exists method
 * Add package#is_built method
+* Add package#build_wait_until_ready method
 
 ### 1.2.0
 * Limit memory consumption when downloading large package file